将部分代码整理到库中
