+++ base/package.json
@@ -6,7 +6,7 @@
   "private": true,
   "main": "dist/handbook.cjs",
   "module": "dist/handbook.js",
-  "types": "dist/index.d.ts",
+  "types": "dist/handbook.d.ts",
   "scripts": {
     "build": "vue-tsc -b && vite build"
   },
@@ -16,6 +16,7 @@
   },
   "devDependencies": {
     "@vitejs/plugin-vue": "^5.1.2",
+    "sass": "^1.77.8",
     "typescript": "^5.5.4",
     "vite": "^5.4.2",
     "vite-plugin-dts": "^4.0.3",
