+++ base/tsconfig.json
@@ -21,5 +21,5 @@
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true
   },
-  "include": ["index.ts", "src/**/*.ts", "src/**/*.vue"]
+  "include": ["handbook.ts", "src/**/*.ts", "Handbook.vue"]
 }
